+++ bittytax/conv/parsers/__init__.py
@@ -14,6 +14,7 @@ from . import electrum
 from . import etherscan
 from . import gatehub
 from . import hitbtc
+from . import kucoin
 from . import ledgerlive
 from . import okex
 from . import poloniex