+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bittrex parser: Adjust quantity for partially filled orders.
 ### Addded
 - HitBTC exchange data files.
+- KuCoin trades data file.
 
 ## Version [0.2.0] Beta (2019-10-30)
 ### Fixed